refactor: 함수 네이밍 의미에 맞게 분리
