+++ src/utils/number.js
@@ -14,19 +14,4 @@ const generateRandomNumber = () => {
   return randomNumber.join("");
 };
 
-const isValidUserInput = (input) => {
-  if (input.length !== EXCEPTION.VALID_NUMBER_LENGTH) return false;
-  if (isNaN(input)) return false;
-  if (input.includes("0")) return false;
-  return true;
-};
-
-const isValidUserAskInput = (input) => {
-  if (input === "1") return true;
-  if (input === "2") return true;
-  return false;
-};
-
 exports.generateRandomNumber = generateRandomNumber;
-exports.isValidUserInput = isValidUserInput;
-exports.isValidUserAskInput = isValidUserAskInput;
